+++ CHANGELOG.md
@@ -6,6 +6,8 @@ title: Changelog
 
 ## Unreleased (????-??-??)
 
+* `<cc-button>`: Add a circle form when in hide-text with an image
+
 ### For devs
 
 * i18n: simplify formatter system and add JSDoc